+++ HISTORY.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Fix encoding unpaired surrogates at start/end of string
+
 1.0.0 / 2016-06-08
 ==================